fix customer delete

+++ js/customer.js
@@ -93,30 +93,26 @@ const updateTableContent = (entity_id) => {
     
 }
 
-console.log(selected_user)
 
-const deleteTableContent = (entity_id) => {
-    console.log(entity_id)
-    customerForm.addEventListener('submit', async function (e) {
-        e.preventDefault();
-        
-        const formData = new FormData(customerForm).entries()
-        console.log(formData)
+const deleteTableContent = async (entity_id) => {
+    console.log(entity_id[0])
+    if (confirm(`Are you sure you want delete ${entity_id[1]} ${entity_id[2]} from the Customers table`)) {
+        // Save it!
         const response = await fetch(`https://planetfitapi.azurewebsites.net/api/customers/${entity_id[0]}`, {
-            method: 'DELETE',
-            headers: { 'Content-Type': 'application/json' },
+            method:'DELETE',
+            headers: { 'Content-Type': 'application/json' }
             //body: JSON.stringify(Object.fromEntries(formData))
         });
+      } 
     
-        const result = await response.json();
-        console.log(result)
+
+
     
-        const modalElement = document.getElementById("customerModal");
-        const modal = bootstrap.Modal.getInstance(modalElement);
-        modal.hide();
-        document.getElementById("addCustomer").reset()
-        
-    });
+    document.getElementById("addCustomer").reset()
+    location.reload()
+
+    
+   
 }
